Move flag to internal/pkg
