+++ cli/info.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"path"
 
-	"github.com/hashicorp/nomad-pack/flag"
 	"github.com/hashicorp/nomad-pack/internal/pkg/cache"
+	"github.com/hashicorp/nomad-pack/internal/pkg/flag"
 	"github.com/hashicorp/nomad-pack/internal/pkg/loader"
 	"github.com/hashicorp/nomad-pack/internal/pkg/variable"
 	"github.com/mitchellh/go-glint"
