+++ cli/main.go
@@ -7,7 +7,7 @@ import (
 	"sort"
 	"text/tabwriter"
 
-	flag "github.com/hashicorp/nomad-pack/flag"
+	flag "github.com/hashicorp/nomad-pack/internal/pkg/flag"
 	"github.com/hashicorp/nomad-pack/internal/pkg/version"
 	"github.com/mitchellh/cli"
 	"github.com/mitchellh/go-glint"
