+++ cli/render.go
@@ -10,9 +10,9 @@ import (
 	"strings"
 
 	v1 "github.com/hashicorp/nomad-openapi/v1"
-	"github.com/hashicorp/nomad-pack/flag"
 	"github.com/hashicorp/nomad-pack/internal/pkg/cache"
 	"github.com/hashicorp/nomad-pack/internal/pkg/errors"
+	"github.com/hashicorp/nomad-pack/internal/pkg/flag"
 	"github.com/hashicorp/nomad-pack/terminal"
 	"github.com/posener/complete"
 )
